requirements.txt
tensorflow==2.8.0
keras==2.8.0
pandas==1.4.2
numpy==1.22.3
streamlit==1.10.0
opencv-python==4.6.0.66
pytorch